Change Types change type of add and subtract from int to double

+++ Calculator/Abstractions/IAdder.cs
@@ -2,6 +2,6 @@
 {
     public interface IAdder
     {
-        int Add(int first,int second);
+        double Add(double first,double second);
     }
 }
